+++ src/ca/goldorion/mcrpcreator/MainApp.java
@@ -1,14 +1,10 @@
 package ca.goldorion.mcrpcreator;
 
-import ca.goldorion.mcrpcreator.models.BlockOutputModel;
-import ca.goldorion.mcrpcreator.models.ProceduralBlockModel;
+import ca.goldorion.mcrpcreator.models.BlockModel;
 import ca.goldorion.mcrpcreator.ui.BlockOverviewController;
-import ca.goldorion.mcrpcreator.ui.BlockTypeSelectorController;
 import ca.goldorion.mcrpcreator.ui.RootLayoutController;
-import ca.goldorion.mcrpcreator.ui.blocks.BlockOutputEditDialogController;
-import ca.goldorion.mcrpcreator.ui.blocks.ProceduralBlockEditDialogController;
-import ca.goldorion.mcrpcreator.ui.blocks.others.ExtensionsOutputEditController;
-import ca.goldorion.mcrpcreator.ui.blocks.others.ExtensionsProceduralEditController;
+import ca.goldorion.mcrpcreator.ui.blocks.BlockEditDialogController;
+import ca.goldorion.mcrpcreator.ui.blocks.others.ExtensionsEditController;
 import javafx.application.Application;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
@@ -23,11 +19,12 @@ import java.io.IOException;
 
 public class MainApp extends Application {
     public static final String appName = "MCR Plugin Creator";
-    public static final String version = "0.3.0";
+    public static final String version = "0.4.0";
 
     private Stage primaryStage;
     private BorderPane rootLayout;
     private final ObservableList blockData = FXCollections.observableArrayList();
+
 
     public static void main(String[] args) {
         launch(args);
@@ -37,8 +34,6 @@ public class MainApp extends Application {
         /*
           To add new templates, write them here
          */
-        blockData.add(new BlockOutputModel("Block Output"));
-
     }
 
     public ObservableList getBlockData(){
@@ -86,7 +81,7 @@ public class MainApp extends Application {
             // Load block overview.
             FXMLLoader loader = new FXMLLoader();
 
-            loader.setLocation(MainApp.class.getResource("/BlockOverview.fxml"));
+            loader.setLocation(MainApp.class.getResource("/ca/goldorion/mcrpcreator/ui/BlockOverview.fxml"));
             AnchorPane blockOverview = loader.load();
 
             // Set block overview into the center of root layout.
@@ -100,23 +95,23 @@ public class MainApp extends Application {
         }
     }
 
-    public boolean showBlockOutputEditDialog(BlockOutputModel blockModel){
+    public boolean showBlockEditDialog(BlockModel blockModel){
         try{
             // Load the fxml file and create a new stage for the popup dialog.
             FXMLLoader loader = new FXMLLoader();
-            loader.setLocation(MainApp.class.getResource("/BlockOutputEditDialog.fxml"));
+            loader.setLocation(MainApp.class.getResource("/ca/goldorion/mcrpcreator/ui/blocks/BlockEditDialog.fxml"));
             AnchorPane page = loader.load();
 
             //Create the Dialog Stage
             Stage dialogStage = new Stage();
-            dialogStage.setTitle("Edit Output Block");
+            dialogStage.setTitle("Edit Procedure Block");
             dialogStage.initModality(Modality.WINDOW_MODAL);
             dialogStage.initOwner(primaryStage);
             Scene scene = new Scene(page);
             dialogStage.setScene(scene);
 
             //Set the block into the controller
-            BlockOutputEditDialogController controller = loader.getController();
+            BlockEditDialogController controller = loader.getController();
             controller.setDialogStage(dialogStage);
             controller.setBlockOutputModel(blockModel);
             controller.setMainApp(this);
@@ -131,34 +126,7 @@ public class MainApp extends Application {
         }
     }
 
-    public void showBlockTypeSelector() {
-        try {
-            // Load the block selector menu window
-            FXMLLoader loader = new FXMLLoader();
-            loader.setLocation(MainApp.class.getResource("/BlockTypeSelector.fxml"));
-            AnchorPane page = loader.load();
-
-            //Create the Dialog Stage
-            Stage dialogStage = new Stage();
-            dialogStage.setTitle("Select Block Type");
-            dialogStage.initModality(Modality.WINDOW_MODAL);
-            dialogStage.initOwner(primaryStage);
-            Scene scene = new Scene(page);
-            dialogStage.setScene(scene);
-
-            //Set the block into the controller
-            BlockTypeSelectorController controller = loader.getController();
-            controller.setDialogStage(dialogStage);
-            controller.setMainApp(this);
-
-            dialogStage.showAndWait();
-
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-    }
-
-    public void showExtensionsOutputEdit(BlockOutputModel blockOutputModel){
+    public void showExtensionsBlockEdit(BlockModel outputBlockModel){
         try {
             // Load the extension edit window
             FXMLLoader loader = new FXMLLoader();
@@ -173,35 +141,9 @@ public class MainApp extends Application {
             Scene scene = new Scene(page);
             dialogStage.setScene(scene);
 
-            ExtensionsOutputEditController controller = loader.getController();
+            ExtensionsEditController controller = loader.getController();
             controller.setExtensionStage(dialogStage);
-            controller.setBlockOutputModel(blockOutputModel);
-            controller.setMainApp(this);
-
-            dialogStage.showAndWait();
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-    }
-
-    public void showExtensionsProceduralEdit(ProceduralBlockModel proceduralBlockModel){
-        try {
-            // Load the extension edit window
-            FXMLLoader loader = new FXMLLoader();
-            loader.setLocation(MainApp.class.getResource("/ExtensionsProceduralEdit.fxml"));
-            AnchorPane page = loader.load();
-
-            //Create the Dialog Stage
-            Stage dialogStage = new Stage();
-            dialogStage.setTitle("Edit Extensions");
-            dialogStage.initModality(Modality.WINDOW_MODAL);
-            dialogStage.initOwner(primaryStage);
-            Scene scene = new Scene(page);
-            dialogStage.setScene(scene);
-
-            ExtensionsProceduralEditController controller = loader.getController();
-            controller.setExtensionStage(dialogStage);
-            controller.setProceduralBlockModel(proceduralBlockModel);
+            controller.setBlockOutputModel(outputBlockModel);
             controller.setMainApp(this);
 
             dialogStage.showAndWait();
@@ -217,34 +159,4 @@ public class MainApp extends Application {
         return primaryStage;
     }
 
-    public boolean showProceduralBlockEditDialog(ProceduralBlockModel blockModel) {
-        try{
-            // Load the fxml file and create a new stage for the popup dialog.
-            FXMLLoader loader = new FXMLLoader();
-            loader.setLocation(MainApp.class.getResource("/ProceduralBlockEditDialog.fxml"));
-            AnchorPane page = loader.load();
-
-            //Create the Dialog Stage
-            Stage dialogStage = new Stage();
-            dialogStage.setTitle("Edit Procedure Block");
-            dialogStage.initModality(Modality.WINDOW_MODAL);
-            dialogStage.initOwner(primaryStage);
-            Scene scene = new Scene(page);
-            dialogStage.setScene(scene);
-
-            //Set the block into the controller
-            ProceduralBlockEditDialogController controller = loader.getController();
-            controller.setDialogStage(dialogStage);
-            controller.setProceduralBlockModel(blockModel);
-            controller.setMainApp(this);
-
-            dialogStage.showAndWait();
-
-            return controller.isOkClicked();
-
-        } catch (IOException e) {
-            e.printStackTrace();
-            return false;
-        }
-    }
 }