+++ src/ca/goldorion/mcrpcreator/io/input/Import.java
@@ -2,7 +2,8 @@ package ca.goldorion.mcrpcreator.io.input;
 
 import ca.goldorion.mcrpcreator.MainApp;
 import ca.goldorion.mcrpcreator.io.jsons.BlockOutput;
-import ca.goldorion.mcrpcreator.models.BlockOutputModel;
+import ca.goldorion.mcrpcreator.io.jsons.ProceduralBlock;
+import ca.goldorion.mcrpcreator.models.BlockModel;
 import ca.goldorion.mcrpcreator.utils.AlertUtils;
 import com.google.gson.Gson;
 
@@ -17,10 +18,13 @@ public class Import {
         Gson gson = new Gson();
         try {
             BlockOutput blockOutput = gson.fromJson(new BufferedReader(new FileReader(file)), BlockOutput.class);
-            BlockOutputModel blockModel = new BlockOutputModel(file.getName().replace(".json", ""));
+            ProceduralBlock proceduralBlock = gson.fromJson(new BufferedReader(new FileReader(file)), ProceduralBlock.class);
+            BlockModel blockModel = new BlockModel(file.getName().replace(".json", ""));
             blockModel.setFileName(file.getName().replace(".json", ""));
             if(blockOutput.getOutput() != null){
-                ca.goldorion.mcrpcreator.io.input.BlockOutput.blockOutput(mainApp, blockModel, blockOutput);
+                BlockOutputImport.blockOutput(mainApp, blockModel, blockOutput);
+            } else if(proceduralBlock.isPreviousStatement() == true) {
+                ProceduralBlockImport.proceduralBlock(mainApp, blockModel, proceduralBlock);
             } else{
                 AlertUtils.error("Block Not supported", "The block you tried to import is not supported. Please import a supported block only.");
             }